sys-fs/squashfs-tools: Modified ebuild adds xz USE flag
See bug #854702. I was trying to rebuild squashfs-utils in order to add xz
decompression support to the unsquashfs utility and I had to look at the ebuild
to find out how to do that. Turns out that xz support is hidden behind the lzma
USE flag (because lzma support _also_ depends on app-arch/xz-utils), which
seems counterintuitive to me.

I'm submitting a modified ebuild for consideration that introduces an xz USE
flag to enable xz support, decoupling it from the lzma USE flag.