sys-boot/grub-2.12_p1-r4: Patch to correct ZFS handling
The grub-mkconfig utility is failing to properly identify the zpool name,
which causes it to produce a grub.cfg file that is non-bootable. Patches
have existed for ages but haven't become part of upstream.

The new _p1 ebuild applies the patch and also makes it distinct from
the current stable ebuild in the Gentoo repo.